Move code to cpp files
